Struct matrix_sdk_crypto::backups::RecoveryKey

pub struct RecoveryKey { /* fields omitted */ }

This is supported on backups_v1 only.

The private part of a backup key.

Implementations

impl RecoveryKey

pub fn new() -> Result<Self, RandomError>

Create a new random recovery key.

pub fn from_bytes(key: [u8; 32]) -> Self

Create a new recovery key from the given byte array.

Warning: You need to make sure that the byte array contains correct random data, either by using a random number generator or by using an exported version of a previously created RecoveryKey.

pub fn from_base64(key: &str) -> Result<Self, DecodeError>

Try to create a RecoveryKey from a base64 export of a RecoveryKey.

pub fn to_base64(&self) -> String

Export the RecoveryKey as a base64 encoded string.

pub fn from_base58(value: &str) -> Result<Self, DecodeError>

Try to create a RecoveryKey from a base58 export of a RecoveryKey.

pub fn to_base58(&self) -> String

Export the RecoveryKey as a base58 encoded string.

pub fn megolm_v1_public_key(&self) -> MegolmV1BackupKey

Extract the megolm.v1 public key from this RecoveryKey.

pub fn pickle(&self, pickle_key: &[u8]) -> PickledRecoveryKey

Export this RecoveryKey as an encrypted pickle that can be safely stored.

pub fn from_pickle(
pickle: PickledRecoveryKey,
pickle_key: &[u8]
) -> Result<Self, UnpicklingError>

Try to import a RecoveryKey from a previously exported pickle.

pub fn decrypt_v1(
&self,
mac: String,
ephemeral_key: String,
ciphertext: String
) -> Result<String, OlmPkDecryptionError>

Try to decrypt the given ciphertext using this RecoveryKey.

This will use the m.megolm_backup.v1.curve25519-aes-sha2 algorithm to decrypt the given ciphertext.