Enum matrix_sdk_crypto::requests::IncomingResponse

pub enum IncomingResponse<'a> {
    KeysUpload(&'a KeysUploadResponse),
    KeysQuery(&'a KeysQueryResponse),
    ToDevice(&'a ToDeviceResponse),
    KeysClaim(&'a KeysClaimResponse),
    SigningKeysUpload(&'a SigningKeysUploadResponse),
    SignatureUpload(&'a SignatureUploadResponse),
    RoomMessage(&'a RoomMessageResponse),
    KeysBackup(&'a KeysBackupResponse),
}

Enum over all the incoming responses we need to receive.

Variants

KeysUpload(&'a KeysUploadResponse)

The keys upload response, notifying us about the amount of uploaded one-time keys.

KeysQuery(&'a KeysQueryResponse)

The keys query response, giving us the device and cross singing keys of other users.

ToDevice(&'a ToDeviceResponse)

The to-device response, an empty response.

KeysClaim(&'a KeysClaimResponse)

The key claiming requests, giving us new one-time keys of other users so new Olm sessions can be created.

SigningKeysUpload(&'a SigningKeysUploadResponse)

The cross signing keys upload response, marking our private cross signing identity as shared.

SignatureUpload(&'a SignatureUploadResponse)

The cross signing signature upload response.

RoomMessage(&'a RoomMessageResponse)

A room message response, usually for interactive verifications.

KeysBackup(&'a KeysBackupResponse)

Response for the server-side room key backup request.