Struct matrix_sdk_crypto::store::SledStore

pub struct SledStore { /* fields omitted */ }

An in-memory only store that will forget all the E2EE key once it’s dropped.

Implementations

impl SledStore

pub fn open_with_passphrase(
path: impl AsRef<Path>,
passphrase: Option<&str>
) -> Result<Self>

Open the sled based cryptostore at the given path using the given passphrase to encrypt private data.

pub fn open_with_database(db: Db, passphrase: Option<&str>) -> Result<Self>

Create a sled based cryptostore using the given sled database. The given passphrase will be used to encrypt private data.