Struct matrix_sdk_crypto::AttachmentDecryptor

pub struct AttachmentDecryptor<'a, R: Read> { /* private fields */ }

A wrapper that transparently encrypts anything that implements Read as an Matrix attachment.

Implementations

impl<'a, R: Read + 'a> AttachmentDecryptor<'a, R>

pub fn new( input: &'a mut R, info: MediaEncryptionInfo ) -> Result<AttachmentDecryptor<'a, R>, DecryptorError>

Wrap the given reader decrypting all the data we read from it.

Arguments

reader - The Reader that should be wrapped and decrypted.

info - The encryption info that is necessary to decrypt data from the reader.

Examples

let data = "Hello world".to_owned();
let mut cursor = Cursor::new(data.clone());

let mut encryptor = AttachmentEncryptor::new(&mut cursor);

let mut encrypted = Vec::new();
encryptor.read_to_end(&mut encrypted).unwrap();
let info = encryptor.finish();

let mut cursor = Cursor::new(encrypted);
let mut decryptor = AttachmentDecryptor::new(&mut cursor, info).unwrap();
let mut decrypted_data = Vec::new();
decryptor.read_to_end(&mut decrypted_data).unwrap();

let decrypted = String::from_utf8(decrypted_data).unwrap();