Struct matrix_sdk_crypto::AttachmentEncryptor

pub struct AttachmentEncryptor<'a, R: Read + ?Sized> { /* private fields */ }

A wrapper that transparently encrypts anything that implements Read.

Implementations

impl<'a, R: Read + ?Sized + 'a> AttachmentEncryptor<'a, R>

pub fn new(reader: &'a mut R) -> Self

Wrap the given reader encrypting all the data we read from it.

After all the reads are done, and all the data is encrypted that we wish to encrypt a call to finish() is necessary to get the decryption key for the data.

Arguments

reader - The Reader that should be wrapped and enrypted.

Panics

Panics if we can’t generate enough random data to create a fresh encryption key.

Examples

let data = "Hello world".to_owned();
let mut cursor = Cursor::new(data.clone());

let mut encryptor = AttachmentEncryptor::new(&mut cursor);

let mut encrypted = Vec::new();
encryptor.read_to_end(&mut encrypted).unwrap();
let key = encryptor.finish();

pub fn finish(self) -> MediaEncryptionInfo

Consume the encryptor and get the encryption key.