Struct matrix_sdk_crypto::Device

pub struct Device { /* private fields */ }

A device represents a E2EE capable client of an user.

Implementations

impl Device

pub async fn start_verification(&self) -> StoreResult<(Sas, ToDeviceRequest)>

Start a interactive verification with this Device

Returns a Sas object and a to-device request that needs to be sent out.

This method has been deprecated in the spec and the request_verification() method should be used instead.

pub fn is_our_own_device(&self) -> bool

Is this our own device?

pub fn is_owner_of_session(&self, session: &InboundGroupSession) -> bool

Does the given InboundGroupSession belong to this device?

An InboundGroupSession is exchanged between devices as an Olm encrypted m.room_key event. This method determines if this Device can be confirmed as the creator and owner of the m.room_key.

pub async fn request_verification(
&self
) -> (VerificationRequest, OutgoingVerificationRequest)

Request an interactive verification with this Device.

Returns a VerificationRequest object and a to-device request that needs to be sent out.

pub async fn request_verification_with_methods(
&self,
methods: Vec<VerificationMethod>
) -> (VerificationRequest, OutgoingVerificationRequest)

Request an interactive verification with this Device.

Returns a VerificationRequest object and a to-device request that needs to be sent out.

Arguments

methods - The verification methods that we want to support.

pub fn is_verified(&self) -> bool

Is this device considered to be verified.

This method returns true if either is_locally_trusted() returns true or if is_cross_signing_trusted() returns true.

pub fn is_cross_signing_trusted(&self) -> bool

Is this device considered to be verified using cross signing.

pub async fn verify(&self) -> Result<SignatureUploadRequest, SignatureError>

Manually verify this device.

This method will attempt to sign the device using our private cross signing key.

This method will always fail if the device belongs to someone else, we can only sign our own devices.

It can also fail if we don’t have the private part of our self-signing key.

Returns a request that needs to be sent out for the device to be marked as verified.

pub async fn set_local_trust(&self, trust_state: LocalTrust) -> StoreResult<()>

Set the local trust state of the device to the given state.

This won’t affect any cross signing trust state, this only sets a flag marking to have the given trust state.

Arguments

trust_state - The new trust state that should be set for the device.

pub async fn encrypt_room_key_for_forwarding(
&self,
session: InboundGroupSession,
message_index: Option<u32>
) -> Result<(Session, Raw<ToDeviceEncryptedEventContent>), OlmError>

Encrypt the given inbound group session as a forwarded room key for this device.