Struct matrix_sdk_crypto::QrVerification

pub struct QrVerification { /* private fields */ }

Available on crate feature qrcode only.

An object controlling QR code style key verification flows.

Implementations

impl QrVerification

pub fn has_been_scanned(&self) -> bool

Has the QR verification been scanned by the other side.

When the verification object is in this state it’s required that the user confirms that the other side has scanned the QR code.

pub fn has_been_confirmed(&self) -> bool

Has the scanning of the QR code been confirmed by us.

pub fn user_id(&self) -> &UserId

Get our own user id.

pub fn other_user_id(&self) -> &UserId

Get the user id of the other user that is participating in this verification flow.

pub fn other_device_id(&self) -> &DeviceId

Get the device ID of the other side.

pub fn we_started(&self) -> bool

Did we initiate the verification request

pub fn cancel_info(&self) -> Option<CancelInfo>

Get info about the cancellation if the verification flow has been cancelled.

pub fn is_done(&self) -> bool

Has the verification flow completed.

pub fn is_cancelled(&self) -> bool

Has the verification flow been cancelled.

pub fn is_self_verification(&self) -> bool

Is this a verification that is veryfying one of our own devices

pub fn reciprocated(&self) -> bool

Have we successfully scanned the QR code and are able to send a reciprocation event.

pub fn flow_id(&self) -> &FlowId

Get the unique ID that identifies this QR code verification flow.

pub fn room_id(&self) -> Option<&RoomId>

Get the room id if the verification is happening inside a room.

pub fn to_qr_code(&self) -> Result<QrCode, EncodingError>

Generate a QR code object that is representing this verification flow.

The QrCode can then be rendered as an image or as an unicode string.

The to_bytes() method can be used to instead output the raw bytes that should be encoded as a QR code.

pub fn to_bytes(&self) -> Result<Vec<u8>, EncodingError>

Generate a the raw bytes that should be encoded as a QR code is representing this verification flow.

The to_qr_code() method can be used to instead output a QrCode object that can be rendered.

pub fn cancel(&self) -> Option<OutgoingVerificationRequest>

Cancel the verification flow.

pub fn cancel_with_code( &self, code: CancelCode ) -> Option<OutgoingVerificationRequest>

Cancel the verification.

This cancels the verification with given CancelCode.

Note: This method should generally not be used, the cancel() method should be preferred. The SDK will automatically cancel with the appropriate cancel code, user initiated cancellations should only cancel with the CancelCode::User

Returns None if the Sas object is already in a canceled state, otherwise it returns a request that needs to be sent out.

pub fn reciprocate(&self) -> Option<OutgoingVerificationRequest>

Notify the other side that we have successfully scanned the QR code and that the QR verification flow can start.

This will return some OutgoingContent if the object is in the correct state to start the verification flow, otherwise None.

pub fn confirm_scanning(&self) -> Option<OutgoingVerificationRequest>

Confirm that the other side has scanned our QR code.

pub fn changes(&self) -> impl Stream<Item = QrVerificationState>

Listen for changes in the QrCode verification process.

The changes are presented as a stream of QrVerificationState values.

pub fn state(&self) -> QrVerificationState

Get the current state the verification process is in.

To listen to changes to the QrVerificationState use the QrVerification::changes method.