Module matrix_sdk_crypto::types

Module containing customized types modeling Matrix keys and events.

These types were mostly taken from the Ruma project. The types differ in two important ways to the Ruma types of the same name:

They are using vodozemac types so we directly deserialize into a vodozemac Curve25519 or Ed25519 key.

They support lossless serialization cycles in a canonical JSON supported way, meaning the white-space and field order won’t be preserved but the data will.

Modules

events

Types modeling end-to-end encryption related Matrix events

Structs

CrossSigningKey

A cross signing key.

DeviceKeys

Identity keys for a device.

InvalidSignature

Represents a signature that could not be decoded.

MegolmV1AuthData

Auth data for the m.megolm_backup.v1.curve25519-aes-sha2 backup algorithm as defined in the spec.

Signatures

Signatures for a signed object.

SignedKey

A key for the SignedCurve25519 algorithm

SigningKeys

A collection of signing keys, a map from the key id to the signing key.

UnsignedDeviceInfo

Additional data added to device key information by intermediate servers.

Enums

DeviceKey

An enum over the different key types a device can have.

EventEncryptionAlgorithm

An encryption algorithm to be used to encrypt messages sent to a room.

OneTimeKey

A one-time public key for “pre-key” messages.

RoomKeyBackupInfo

Information pertaining to a room key backup. Can be used to upload a new backup version as defined in the spec.

Signature

Represents a potentially decoded signature (but not a validated one).

SigningKey

An enum over the different key types a cross-signing key can have.