Module matrix_sdk_crypto::types::events

Types modeling end-to-end encryption related Matrix events

These types aim to provide a more strict variant of the equivalent Ruma types. Once deserialized they aim to zeroize all the secret material once the type is dropped.

Modules

forwarded_room_key

Types for m.forwarded_room_key to-device events.

olm_v1

Module containing specialized event types that were decrypted using the Olm protocol

room

Types for room events.

room_key

Types for m.room_key to-device events.

room_key_request

Types for the m.room_key_request events.

secret_send

Types for m.secret.send to-device events.

Structs

ToDeviceCustomEvent

A to-device event with an unknown type and content.

ToDeviceEvent

Generic to-device event with a known type and content.

Enums

ToDeviceEvents

An enum over the various to-device events we support.

Traits

EventType

A trait for event contents to define their event type.