Module matrix_sdk_crypto::types::events::olm_v1

Module containing specialized event types that were decrypted using the Olm protocol

Structs

DecryptedOlmV1Event

An m.olm.v1.curve25519-aes-sha2 decrypted to-device event.

OlmV1Keys

Public keys used for an m.olm.v1.curve25519-aes-sha2 event.

ToDeviceCustomEvent

A decrypted to-device event with an unknown type and content.

Enums

AnyDecryptedOlmEvent

An enum over the various events that were decrypted using the m.olm.v1.curve25519-aes-sha2 algorithm.

Type Definitions

DecryptedForwardedRoomKeyEvent

An m.forwarded_room_key event that was decrypted using the m.olm.v1.curve25519-aes-sha2 algorithm

DecryptedRoomKeyEvent

An m.room_key event that was decrypted using the m.olm.v1.curve25519-aes-sha2 algorithm

DecryptedSecretSendEvent

An m.secret.send event that was decrypted using the m.olm.v1.curve25519-aes-sha2 algorithm