Type Definition matrix_sdk_crypto::types::events::olm_v1::DecryptedForwardedRoomKeyEvent

pub type DecryptedForwardedRoomKeyEvent = DecryptedOlmV1Event<ForwardedRoomKeyContent>;

An m.forwarded_room_key event that was decrypted using the m.olm.v1.curve25519-aes-sha2 algorithm

Implementations

impl DecryptedForwardedRoomKeyEvent

pub fn room_key_info(&self) -> Option<SupportedKeyInfo>

Get the unique info about the room key that is contained in this forwarded room key event.

Returns None if we do not understand the algorithm that was used to encrypt the event.