Type Alias matrix_sdk_crypto::types::events::room::encrypted::EncryptedEvent

pub type EncryptedEvent = Event<RoomEncryptedEventContent>;

An m.room.encrypted room event.

Aliased Type

struct EncryptedEvent {
    pub sender: OwnedUserId,
    pub event_id: OwnedEventId,
    pub content: RoomEncryptedEventContent,
    pub origin_server_ts: MilliSecondsSinceUnixEpoch,
    pub unsigned: BTreeMap<String, Value>,
    /* private fields */
}

Fields
sender: OwnedUserId

Contains the fully-qualified ID of the user who sent this event.

event_id: OwnedEventId

The globally unique identifier for this event.

content: RoomEncryptedEventContent

The body of this event, as created by the client which sent it.

origin_server_ts: MilliSecondsSinceUnixEpoch

Timestamp (in milliseconds since the unix epoch) on originating homeserver when this event was sent.

unsigned: BTreeMap<String, Value>

Contains optional extra information about the event.

Implementations

impl EncryptedEvent

pub fn room_key_info(&self, room_id: &RoomId) -> Option<SupportedKeyInfo>

Get the unique info about the room key that was used to encrypt this event.

Returns None if we do not understand the algorithm that was used to encrypt the event.