Module matrix_sdk_crypto::types::events::room_key_request

Types for the m.room_key_request events.

Structs

MegolmV1AesSha2Content

The content for a m.megolm.v2.aes-sha2 m.room_key_request event.

RoomKeyRequestContent

The content for a m.room_key_request event.

UnknownRoomKeyRequest

An unknown and unsupported m.room_key_request algorithm.

Enums

Action

Enum describing different actions a room key request event can be carrying.

RequestedKeyInfo

Info about the room key that is being requested.

SupportedKeyInfo

Info about the room key that is being requested.

Type Definitions

RoomKeyRequestEvent

The m.room_key_request to-device event.