Trait matrix_sdk_crypto::types::events::EventType

pub trait EventType {
    const EVENT_TYPE: &'static str;

    fn event_type(&self) -> &'static str { ... }
}

A trait for event contents to define their event type.

Required Associated Constants

const EVENT_TYPE: &'static str

The event type of the event content.

Provided Methods

fn event_type(&self) -> &'static str

Get the event type of the event content.

Note: This should never be implemented manually, this takes the event type from the constant.