Module matrix_sdk_crypto::olm

The crypto specific Olm objects.

Note: You’ll only be interested in these if you are implementing a custom CryptoStore.

Structs

BackedUpRoomKey

A backed up version of an InboundGroupSession

CrossSigningStatus

Struct representing the state of our private cross signing keys, it shows which private cross signing keys we have locally stored.

Curve25519PublicKey

Struct representing a Curve25519 public key.

EncryptionSettings

Settings for an encrypted room.

ExportedRoomKey

An exported version of an InboundGroupSession

IdentityKeys

Struct holding the two public identity keys of an Account.

InboundGroupSession

A structure representing an inbound group session.

OlmMessageHash

A hash of a successfully decrypted Olm message.

OutboundGroupSession

Outbound group session.

PickledAccount

A pickled version of an Account.

PickledCrossSigningIdentity

The pickled version of a PrivateCrossSigningIdentity.

PickledInboundGroupSession

A pickled version of an InboundGroupSession.

PickledOutboundGroupSession

A pickled version of an InboundGroupSession.

PickledSession

A pickled version of a Session.

PrivateCrossSigningIdentity

Private cross signing identity.

ReadOnlyAccount

Account holding identity keys for which sessions can be created.

Session

Cryptographic session that enables secure communication between two Accounts

SessionKey

The session key, can be used to create a InboundGroupSession.

Enums

SessionCreationError

An error type for the creation of group sessions.

SessionExportError

An error type for the export of inbound group sessions.

ShareInfo

Struct holding info about the share state of a outbound group session.