Struct matrix_sdk_crypto::olm::InboundGroupSession

pub struct InboundGroupSession {
    pub room_id: Arc<RoomId>,
    /* private fields */
}

A structure representing an inbound group session.

Inbound group sessions, also known as “room keys”, are used to facilitate the exchange of room messages among a group of participants. The inbound variant of the group session is used to decrypt the room messages.

This struct wraps the vodozemac type of the same name, and adds additional Matrix-specific data to it. Additionally, the wrapper ensures thread-safe access of the vodozemac type.

Fields
room_id: Arc<RoomId>

The Room this GroupSession belongs to

Implementations

impl InboundGroupSession

pub fn new( sender_key: Curve25519PublicKey, signing_key: Ed25519PublicKey, room_id: &RoomId, session_key: &SessionKey, encryption_algorithm: EventEncryptionAlgorithm, history_visibility: Option<HistoryVisibility> ) -> Result<Self, SessionCreationError>

Create a new inbound group session for the given room.

These sessions are used to decrypt room messages.

Arguments

sender_key - The public Curve25519 key of the account that sent us the session.

signing_key - The public Ed25519 key of the account that sent us the session.

room_id - The id of the room that the session is used in.

session_key - The private session key that is used to decrypt messages.

pub fn from_export( exported_session: &ExportedRoomKey ) -> Result<Self, SessionCreationError>

Create a InboundGroupSession from an exported version of the group session.

Most notably this can be called with an ExportedRoomKey from a previous export() call.

pub async fn pickle(&self) -> PickledInboundGroupSession

Store the group session as a base64 encoded string.

Arguments

pickle_mode - The mode that was used to pickle the group session, either an unencrypted mode or an encrypted using passphrase.

pub async fn export(&self) -> ExportedRoomKey

Export this session at the first known message index.

If only a limited part of this session should be exported use export_at_index().

pub fn sender_key(&self) -> Curve25519PublicKey

Get the sender key that this session was received from.

pub fn backed_up(&self) -> bool

Has the session been backed up to the server.

pub fn reset_backup_state(&self)

Reset the backup state of the inbound group session.

pub fn mark_as_backed_up(&self)

For testing, allow to manually mark this GroupSession to have been backed up

pub fn signing_keys(&self) -> &SigningKeys<DeviceKeyAlgorithm>

Get the map of signing keys this session was received from.

pub async fn export_at_index(&self, message_index: u32) -> ExportedRoomKey

Export this session at the given message index.

pub fn from_pickle( pickle: PickledInboundGroupSession ) -> Result<Self, PickleError>

Restore a Session from a previously pickled string.

Returns the restored group session or a UnpicklingError if there was an error.

Arguments

pickle - The pickled version of the InboundGroupSession.

pickle_mode - The mode that was used to pickle the session, either an unencrypted mode or an encrypted using passphrase.

pub fn room_id(&self) -> &RoomId

The room where this session is used in.

pub fn session_id(&self) -> &str

Returns the unique identifier for this session.

pub fn algorithm(&self) -> &EventEncryptionAlgorithm

The algorithm that this inbound group session is using to decrypt events.

pub fn first_known_index(&self) -> u32

Get the first message index we know how to decrypt.

pub fn has_been_imported(&self) -> bool

Has the session been imported from a file or server-side backup? As opposed to being directly received as an m.room_key event.

pub async fn compare(&self, other: &InboundGroupSession) -> SessionOrdering

Check if the InboundGroupSession is better than the given other InboundGroupSession

pub async fn decrypt( &self, event: &EncryptedEvent ) -> Result<(Raw<AnyTimelineEvent>, u32), MegolmError>

Decrypt an event from a room timeline.

Arguments

event - The event that should be decrypted.