Module matrix_sdk_crypto::requests

Modules containing customized request types.

Structs

KeysBackupRequest

A request that will back up a batch of room keys to the server.

KeysQueryRequest

Customized version of ruma_client_api::keys::get_keys::v3::Request, without any references.

OutgoingRequest

Outgoing request type, holds the unique ID of the request and the actual request.

RoomMessageRequest

Customized owned request type for sending out room messages.

ToDeviceRequest

Customized version of ruma_client_api::to_device::send_event_to_device::v3::Request

UploadSigningKeysRequest

Request that will publish a cross signing identity.

Enums

IncomingResponse

Enum over all the incoming responses we need to receive.

OutgoingRequests

Enum over the different outgoing requests we can have.

OutgoingVerificationRequest

An enum over the different outgoing verification based requests.