Module matrix_sdk_crypto::store

Types and traits to implement the storage layer for the OlmMachine

The storage layer for the OlmMachine can be customized using a trait. Implementing your own CryptoStore

An in-memory only store is provided as well as a Sled based one, depending on your needs and targets a custom store may be implemented, e.g. for wasm-unknown-unknown an indexeddb store would be needed

let store = Box::new(MemoryStore::new());

let machine = OlmMachine::new_with_store(user_id, device_id, store);

Modules

caches

Collection of small in-memory stores that can be used to cache Olm objects.

Structs

Changes

DeviceChanges

EncryptedPickleKey

An encrypted version of our pickle key, this can be safely stored in a database.

IdentityChanges

MemoryStore

An in-memory only store that will forget all the E2EE key once it’s dropped.

PickleKey

A pickle key that will be used to encrypt all the private keys for Olm.

SledStore

An in-memory only store that will forget all the E2EE key once it’s dropped.

Enums

CryptoStoreError

The crypto store’s error type.

Traits

CryptoStore

Trait abstracting a store that the OlmMachine uses to store cryptographic keys.

Type Definitions

Result

A CryptoStore specific result type.