Struct matrix_sdk_crypto::store::Store

pub struct Store { /* private fields */ }

A wrapper for our CryptoStore trait object.

This is needed because we want to have a generic interface so we can store/restore objects that we can serialize. Since trait objects and generics don’t mix let the CryptoStore store strings and this wrapper adds the generic interface on top.

Implementations

impl Store

pub fn new(
user_id: Arc<UserId>,
identity: Arc<Mutex<PrivateCrossSigningIdentity>>,
store: Arc<dyn CryptoStore>,
verification_machine: VerificationMachine
) -> Self

Create a new Store

pub fn user_id(&self) -> &UserId

UserId associated with this store

pub fn device_id(&self) -> &DeviceId

DeviceId associated with this store

pub fn account(&self) -> &ReadOnlyAccount

The Account associated with this store

pub fn private_identity(&self) -> Arc<Mutex<PrivateCrossSigningIdentity>>

PrivateCrossSigningIdentity associated with this store

pub async fn save_sessions(&self, sessions: &[Session]) -> Result<()>

Save the given Sessions to the store

pub async fn device_display_name(
&self
) -> Result<Option<String>, CryptoStoreError>

Get the display name of our own device.

pub async fn get_readonly_device(
&self,
user_id: &UserId,
device_id: &DeviceId
) -> Result<Option<ReadOnlyDevice>>

Get the read-only device associated with device_id for user_id

pub async fn get_readonly_devices_filtered(
&self,
user_id: &UserId
) -> Result<HashMap<OwnedDeviceId, ReadOnlyDevice>>

Get the read-only version of all the devices that the given user has.

Note: This doesn’t return our own device.

pub async fn get_readonly_devices_unfiltered(
&self,
user_id: &UserId
) -> Result<HashMap<OwnedDeviceId, ReadOnlyDevice>>

Get the read-only version of all the devices that the given user has.

Note: This does also return our own device.

pub async fn get_device_from_curve_key(
&self,
user_id: &UserId,
curve_key: &str
) -> Result<Option<Device>>

Get a device for the given user with the given curve25519 key.

Note: This doesn’t return our own device.

pub async fn get_user_devices_filtered(
&self,
user_id: &UserId
) -> Result<UserDevices>

Get all devices associated with the given user_id

Note: This doesn’t return our own device.

pub async fn get_user_devices(&self, user_id: &UserId) -> Result<UserDevices>

Get all devices associated with the given user_id

Note: This does also return our own device.

pub async fn get_device(
&self,
user_id: &UserId,
device_id: &DeviceId
) -> Result<Option<Device>>

Get a Device copy associated with device_id for user_id

pub async fn get_identity(
&self,
user_id: &UserId
) -> Result<Option<UserIdentities>>

Get the Identity of user_id

pub async fn export_secret(&self, secret_name: &SecretName) -> Option<String>

Try to export the secret with the given secret name.

The exported secret will be encoded as unpadded base64. Returns Null if the secret can’t be found.

Arguments

secret_name - The name of the secret that should be exported.

pub async fn import_cross_signing_keys(
&self,
export: CrossSigningKeyExport
) -> Result<CrossSigningStatus, SecretImportError>

Import the Cross Signing Keys

pub async fn import_secret(
&self,
secret_name: &SecretName,
secret: String
) -> Result<(), SecretImportError>

Import the given secret named secret_name into the keystore.