Trait matrix_sdk_crypto::store::CryptoStore

pub trait CryptoStore: AsyncTraitDeps {
#[must_use]    fn load_account<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<ReadOnlyAccount>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn save_account<'life0, 'async_trait>(
        &'life0 self, 
        account: ReadOnlyAccount
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn load_identity<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<PrivateCrossSigningIdentity>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn save_changes<'life0, 'async_trait>(
        &'life0 self, 
        changes: Changes
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_sessions<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        sender_key: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<Arc<Mutex<Vec<Session>>>>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_inbound_group_session<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId, 
        sender_key: &'life2 str, 
        session_id: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<InboundGroupSession>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_inbound_group_sessions<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<InboundGroupSession>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_outbound_group_sessions<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId
    ) -> Pin<Box<dyn Future<Output = Result<Option<OutboundGroupSession>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn is_user_tracked(&self, user_id: &UserId) -> bool;

    fn has_users_for_key_query(&self) -> bool;

    fn users_for_key_query(&self) -> HashSet<UserId>;

#[must_use]    fn update_tracked_user<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        user: &'life1 UserId, 
        dirty: bool
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_device<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        user_id: &'life1 UserId, 
        device_id: &'life2 DeviceId
    ) -> Pin<Box<dyn Future<Output = Result<Option<ReadOnlyDevice>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_user_devices<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        user_id: &'life1 UserId
    ) -> Pin<Box<dyn Future<Output = Result<HashMap<DeviceIdBox, ReadOnlyDevice>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_user_identity<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        user_id: &'life1 UserId
    ) -> Pin<Box<dyn Future<Output = Result<Option<UserIdentities>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn save_value<'life0, 'async_trait>(
        &'life0 self, 
        key: String, 
        value: String
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn remove_value<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        key: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_value<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        key: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn is_message_known<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        message_hash: &'life1 OlmMessageHash
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

Trait abstracting a store that the OlmMachine uses to store cryptographic keys.

Required methods

#[must_use]fn load_account<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<ReadOnlyAccount>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Load an account that was previously stored.

#[must_use]fn save_account<'life0, 'async_trait>(
&'life0 self,
account: ReadOnlyAccount
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Save the given account in the store.

Arguments

account - The account that should be stored.

#[must_use]fn load_identity<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<PrivateCrossSigningIdentity>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Try to load a private cross signing identity, if one is stored.

#[must_use]fn save_changes<'life0, 'async_trait>(
&'life0 self,
changes: Changes
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Save the set of changes to the store.

Arguments

changes - The set of changes that should be stored.

#[must_use]fn get_sessions<'life0, 'life1, 'async_trait>(
&'life0 self,
sender_key: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Option<Arc<Mutex<Vec<Session>>>>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get all the sessions that belong to the given sender key.

Arguments

sender_key - The sender key that was used to establish the sessions.

#[must_use]fn get_inbound_group_session<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId,
sender_key: &'life2 str,
session_id: &'life3 str
) -> Pin<Box<dyn Future<Output = Result<Option<InboundGroupSession>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Get the inbound group session from our store.

Arguments

room_id - The room id of the room that the session belongs to.

sender_key - The sender key that sent us the session.

session_id - The unique id of the session.

#[must_use]fn get_inbound_group_sessions<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<InboundGroupSession>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get all the inbound group sessions we have stored.

#[must_use]fn get_outbound_group_sessions<'life0, 'life1, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId
) -> Pin<Box<dyn Future<Output = Result<Option<OutboundGroupSession>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get the outobund group sessions we have stored that is used for the given room.

fn is_user_tracked(&self, user_id: &UserId) -> bool

Is the given user already tracked.

fn has_users_for_key_query(&self) -> bool

Are there any tracked users that are marked as dirty.

fn users_for_key_query(&self) -> HashSet<UserId>

Set of users that we need to query keys for. This is a subset of the tracked users.

#[must_use]fn update_tracked_user<'life0, 'life1, 'async_trait>(
&'life0 self,
user: &'life1 UserId,
dirty: bool
) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Add an user for tracking.

Returns true if the user wasn’t already tracked, false otherwise.

Arguments

user - The user that should be marked as tracked.

dirty - Should the user be also marked for a key query.

#[must_use]fn get_device<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId,
device_id: &'life2 DeviceId
) -> Pin<Box<dyn Future<Output = Result<Option<ReadOnlyDevice>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Get the device for the given user with the given device id.

Arguments

user_id - The user that the device belongs to.

device_id - The unique id of the device.

#[must_use]fn get_user_devices<'life0, 'life1, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId
) -> Pin<Box<dyn Future<Output = Result<HashMap<DeviceIdBox, ReadOnlyDevice>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get all the devices of the given user.

Arguments

user_id - The user for which we should get all the devices.

#[must_use]fn get_user_identity<'life0, 'life1, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId
) -> Pin<Box<dyn Future<Output = Result<Option<UserIdentities>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get the user identity that is attached to the given user id.

Arguments

user_id - The user for which we should get the identity.

#[must_use]fn save_value<'life0, 'async_trait>(
&'life0 self,
key: String,
value: String
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Save a serializeable object in the store.

#[must_use]fn remove_value<'life0, 'life1, 'async_trait>(
&'life0 self,
key: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Remove a value from the store.

#[must_use]fn get_value<'life0, 'life1, 'async_trait>(
&'life0 self,
key: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Load a serializeable object from the store.

#[must_use]fn is_message_known<'life0, 'life1, 'async_trait>(
&'life0 self,
message_hash: &'life1 OlmMessageHash
) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Check if a hash for an Olm message stored in the database.

Loading content...