Struct matrix_sdk_crypto::ReadOnlyDevice

pub struct ReadOnlyDevice { /* private fields */ }

A read-only version of a Device.

Implementations

impl ReadOnlyDevice

pub fn new(device_keys: DeviceKeys, trust_state: LocalTrust) -> Self

Create a new Device, this constructor skips signature verification of the keys, TryFrom should be used for completely new devices we receive.

pub fn user_id(&self) -> &UserId

The user id of the device owner.

pub fn device_id(&self) -> &DeviceId

The unique ID of the device.

pub fn display_name(&self) -> Option<&str>

Get the human readable name of the device.

pub fn get_key(&self, algorithm: DeviceKeyAlgorithm) -> Option<&DeviceKey>

Get the key of the given key algorithm belonging to this device.

pub fn curve25519_key(&self) -> Option<Curve25519PublicKey>

Get the Curve25519 key of the given device.

pub fn ed25519_key(&self) -> Option<Ed25519PublicKey>

Get the Ed25519 key of the given device.

pub fn keys(&self) -> &BTreeMap<OwnedDeviceKeyId, DeviceKey>

Get a map containing all the device keys.

pub fn signatures(&self) -> &Signatures

Get a map containing all the device signatures.

pub fn local_trust_state(&self) -> LocalTrust

Get the trust state of the device.

pub fn is_locally_trusted(&self) -> bool

Is the device locally marked as trusted.

pub fn is_blacklisted(&self) -> bool

Is the device locally marked as blacklisted.

Blacklisted devices won’t receive any group sessions.

pub fn algorithms(&self) -> &[EventEncryptionAlgorithm]

Get the list of algorithms this device supports.

pub fn supports_olm(&self) -> bool

Does this device support any of our known Olm encryption algorithms.

pub fn olm_session_config(&self) -> SessionConfig

Get the optimal SessionConfig for this device.

pub fn is_deleted(&self) -> bool

Is the device deleted.

pub async fn from_account(account: &ReadOnlyAccount) -> ReadOnlyDevice

Create a ReadOnlyDevice from an Account

We will have our own device in the store once we receive a keys/query response, but this is useful to create it before we receive such a response.

It also makes it easier to check that the server doesn’t lie about our own device.

Don’t use this after we received a keys/query response, other users/devices might add signatures to our own device, which can’t be replicated locally.