Struct matrix_sdk_crypto::Sas

pub struct Sas { /* fields omitted */ }

Short authentication string object.

Implementations

impl Sas

pub fn user_id(&self) -> &UserId

Get our own user id.

pub fn device_id(&self) -> &DeviceId

Get our own device id.

pub fn other_user_id(&self) -> &UserId

Get the user id of the other side.

pub fn other_device_id(&self) -> &DeviceId

Get the device id of the other side.

pub fn other_device(&self) -> ReadOnlyDevice

Get the device of the other user.

pub fn flow_id(&self) -> &FlowId

Get the unique ID that identifies this SAS verification flow.

pub fn accept(&self) -> Option<OutgoingVerificationRequest>

Accept the SAS verification.

This does nothing if the verification was already accepted, otherwise it returns an AcceptEventContent that needs to be sent out.

pub async fn confirm(
&self
) -> Result<(Option<OutgoingVerificationRequest>, Option<SignatureUploadRequest>), CryptoStoreError>

Confirm the Sas verification.

This confirms that the short auth strings match on both sides.

Does nothing if we’re not in a state where we can confirm the short auth string, otherwise returns a MacEventContent that needs to be sent to the server.

pub fn cancel(&self) -> Option<OutgoingVerificationRequest>

Cancel the verification.

This cancels the verification with the CancelCode::User.

Returns None if the Sas object is already in a canceled state, otherwise it returns a request that needs to be sent out.

pub fn timed_out(&self) -> bool

Has the SAS verification flow timed out.

pub fn can_be_presented(&self) -> bool

Are we in a state where we can show the short auth string.

pub fn is_done(&self) -> bool

Is the SAS flow done.

pub fn is_canceled(&self) -> bool

Is the SAS flow canceled.

pub fn emoji(&self) -> Option<Vec<(&'static str, &'static str)>>

Get the emoji version of the short auth string.

Returns None if we can’t yet present the short auth string, otherwise a Vec of tuples with the emoji and description.

pub fn decimals(&self) -> Option<(u16, u16, u16)>

Get the decimal version of the short auth string.

Returns None if we can’t yet present the short auth string, otherwise a tuple containing three 4-digit integers that represent the short auth string.