Struct matrix_sdk_crypto::VerificationRequest

pub struct VerificationRequest { /* private fields */ }

An object controlling key verification requests.

Interactive verification flows usually start with a verification request, this object lets you send and reply to such a verification request.

After the initial handshake the verification flow transitions into one of the verification methods.

Implementations

impl VerificationRequest

pub fn request( own_user_id: &UserId, own_device_id: &DeviceId, other_user_id: &UserId, methods: Option<Vec<VerificationMethod>> ) -> KeyVerificationRequestEventContent

Create an event content that can be sent as a room event to request verification from the other side. This should be used only for verifications of other users and it should be sent to a room we consider to be a DM with the other user.

pub fn own_user_id(&self) -> &UserId

Our own user id.

pub fn other_user(&self) -> &UserId

The id of the other user that is participating in this verification request.

pub fn other_device_id(&self) -> Option<OwnedDeviceId>

The id of the other device that is participating in this verification.

pub fn room_id(&self) -> Option<&RoomId>

Get the room id if the verification is happening inside a room.

pub fn cancel_info(&self) -> Option<CancelInfo>

Get info about the cancellation if the verification request has been cancelled.

pub fn is_passive(&self) -> bool

Has the verification request been answered by another device.

pub fn is_ready(&self) -> bool

Is the verification request ready to start a verification flow.

pub fn timed_out(&self) -> bool

Has the verification flow timed out.

pub fn their_supported_methods(&self) -> Option<Vec<VerificationMethod>>

Get the supported verification methods of the other side.

Will be present only if the other side requested the verification or if we’re in the ready state.

pub fn our_supported_methods(&self) -> Option<Vec<VerificationMethod>>

Get our own supported verification methods that we advertised.

Will be present only we requested the verification or if we’re in the ready state.

pub fn flow_id(&self) -> &FlowId

Get the unique ID of this verification request

pub fn is_self_verification(&self) -> bool

Is this a verification that is veryfying one of our own devices

pub fn we_started(&self) -> bool

Did we initiate the verification request

pub fn is_done(&self) -> bool

Has the verification flow that was started with this request finished.

pub fn is_cancelled(&self) -> bool

Has the verification flow that was started with this request been cancelled.

pub async fn generate_qr_code( &self ) -> Result<Option<QrVerification>, CryptoStoreError>

Available on crate feature qrcode only.

Generate a QR code that can be used by another client to start a QR code based verification.

pub async fn scan_qr_code( &self, data: QrVerificationData ) -> Result<Option<QrVerification>, ScanError>

Available on crate feature qrcode only.

Start a QR code verification by providing a scanned QR code for this verification flow.

Returns a ScanError if the QR code isn’t valid, None if the verification request isn’t in the ready state or we don’t support QR code verification, otherwise a newly created QrVerification object which will be used for the remainder of the verification flow.

pub fn accept_with_methods( &self, methods: Vec<VerificationMethod> ) -> Option<OutgoingVerificationRequest>

Accept the verification request signaling that our client supports the given verification methods.

Arguments

methods - The methods that we should advertise as supported by us.

pub fn accept(&self) -> Option<OutgoingVerificationRequest>

Accept the verification request.

This method will accept the request and signal that it supports the m.sas.v1, the m.qr_code.show.v1, and m.reciprocate.v1 method.

m.qr_code.show.v1 will only be signaled if the qrcode feature is enabled. This feature is disabled by default. If it’s enabled and QR code scanning should be supported or QR code showing shouldn’t be supported the accept_with_methods() method should be used instead.

pub fn cancel(&self) -> Option<OutgoingVerificationRequest>

Cancel the verification request

pub async fn start_sas( &self ) -> Result<Option<(Sas, OutgoingVerificationRequest)>, CryptoStoreError>

Transition from this verification request into a SAS verification flow.

pub fn changes(&self) -> impl Stream<Item = VerificationRequestState>

Listen for changes in the verification request.

The changes are presented as a stream of VerificationRequestState values.

pub fn state(&self) -> VerificationRequestState

Get the current state the verification request is in.

To listen to changes to the VerificationRequestState use the VerificationRequest::changes method.