Module matrix_sdk_crypto::types

Module containing customized types modeling Matrix keys.

These types were mostly taken from the Ruma project. The types differ in two important ways to the Ruma types of the same name:

They are using vodozemac types so we directly deserialize into a vodozemac curve25519 or ed25519 key.

They support lossless serialization cycles in a canonical JSON supported way, meaning the white-space and field order won’t be preserved but the data will.

Structs

CrossSigningKey

A cross signing key.

DeviceKeys

Identity keys for a device.

SignedKey

A key for the SignedCurve25519 algorithm

UnsignedDeviceInfo

Additional data added to device key information by intermediate servers.

Enums

DeviceKey

An enum over the different key types a device can have.

OneTimeKey

A one-time public key for “pre-key” messages.

SigningKey

An enum over the different key types a cross-signing key can have.

Type Definitions

CrossSigningKeySignatures

Signatures for a CrossSigningKey object.

SignedKeySignatures

Signatures for a SignedKey object.